OtherUserProfile: refactor OtherUserPage page.
